Reject promises with proper errors.
If a request to Keycloak does not get the expected HTTP status, instead
of rejecting the promise with the response body (which could be empty),
reject it with an HttpError.

The HttpError will include the response code, as well as the response
body.

The benefit of rejecting promises with a proper Error is to be able to
get a stack trace when an error occurs.